+++ src/pkcs1v15.rs
@@ -24,9 +24,9 @@ use signature::{
     DigestSigner, DigestVerifier, Keypair, RandomizedDigestSigner, RandomizedSigner,
     SignatureEncoding, Signer, Verifier,
 };
-use subtle::{Choice, ConditionallySelectable, ConstantTimeEq};
 use zeroize::Zeroizing;
 
+use crate::algorithms::pkcs1v15::*;
 use crate::dummy_rng::DummyRng;
 use crate::errors::{Error, Result};
 use crate::internals::{uint_to_be_pad, uint_to_zeroizing_be_pad};
@@ -80,7 +80,7 @@ impl Pkcs1v15Sign {
     {
         Self {
             hash_len: Some(<D as Digest>::output_size()),
-            prefix: generate_prefix::<D>().into_boxed_slice(),
+            prefix: pkcs1v15_generate_prefix::<D>().into_boxed_slice(),
         }
     }
 
@@ -193,25 +193,14 @@ impl Display for Signature {
 /// scheme from PKCS#1 v1.5.  The message must be no longer than the
 /// length of the public modulus minus 11 bytes.
 #[inline]
-pub(crate) fn encrypt<R: CryptoRngCore + ?Sized>(
+fn encrypt<R: CryptoRngCore + ?Sized>(
     rng: &mut R,
     pub_key: &RsaPublicKey,
     msg: &[u8],
 ) -> Result<Vec<u8>> {
     key::check_public(pub_key)?;
 
-    let k = pub_key.size();
-    if msg.len() > k - 11 {
-        return Err(Error::MessageTooLong);
-    }
-
-    // EM = 0x00 || 0x02 || PS || 0x00 || M
-    let mut em = Zeroizing::new(vec![0u8; k]);
-    em[1] = 2;
-    non_zero_random_bytes(rng, &mut em[2..k - msg.len() - 1]);
-    em[k - msg.len() - 1] = 0;
-    em[k - msg.len()..].copy_from_slice(msg);
-
+    let em = pkcs1v15_encrypt_pad(rng, msg, pub_key.size())?;
     let int = Zeroizing::new(BigUint::from_bytes_be(&em));
     uint_to_be_pad(pub_key.raw_int_encryption_primitive(&int)?, pub_key.size())
 }
@@ -226,19 +215,17 @@ pub(crate) fn encrypt<R: CryptoRngCore + ?Sized>(
 /// forge signatures as if they had the private key. See
 /// `decrypt_session_key` for a way of solving this problem.
 #[inline]
-pub(crate) fn decrypt<R: CryptoRngCore + ?Sized>(
+fn decrypt<R: CryptoRngCore + ?Sized>(
     rng: Option<&mut R>,
     priv_key: &RsaPrivateKey,
     ciphertext: &[u8],
 ) -> Result<Vec<u8>> {
     key::check_public(priv_key)?;
 
-    let (valid, out, index) = decrypt_inner(rng, priv_key, ciphertext)?;
-    if valid == 0 {
-        return Err(Error::Decryption);
-    }
+    let em = priv_key.raw_int_decryption_primitive(rng, &BigUint::from_bytes_be(ciphertext))?;
+    let em = uint_to_zeroizing_be_pad(em, priv_key.size())?;
 
-    Ok(out[index as usize..].to_vec())
+    pkcs1v15_encrypt_unpad(em, priv_key.size())
 }
 
 /// Calculates the signature of hashed using
@@ -255,26 +242,13 @@ pub(crate) fn decrypt<R: CryptoRngCore + ?Sized>(
 /// messages to signatures and identify the signed messages. As ever,
 /// signatures provide authenticity, not confidentiality.
 #[inline]
-pub(crate) fn sign<R: CryptoRngCore + ?Sized>(
+fn sign<R: CryptoRngCore + ?Sized>(
     rng: Option<&mut R>,
     priv_key: &RsaPrivateKey,
     prefix: &[u8],
     hashed: &[u8],
 ) -> Result<Vec<u8>> {
-    let hash_len = hashed.len();
-    let t_len = prefix.len() + hashed.len();
-    let k = priv_key.size();
-    if k < t_len + 11 {
-        return Err(Error::MessageTooLong);
-    }
-
-    // EM = 0x00 || 0x01 || PS || 0x00 || T
-    let mut em = vec![0xff; k];
-    em[0] = 0;
-    em[1] = 1;
-    em[k - t_len - 1] = 0;
-    em[k - t_len..k - hash_len].copy_from_slice(prefix);
-    em[k - hash_len..k].copy_from_slice(hashed);
+    let em = pkcs1v15_sign_pad(prefix, hashed, priv_key.size())?;
 
     uint_to_zeroizing_be_pad(
         priv_key.raw_int_decryption_primitive(rng, &BigUint::from_bytes_be(&em))?,
@@ -284,125 +258,10 @@ pub(crate) fn sign<R: CryptoRngCore + ?Sized>(
 
 /// Verifies an RSA PKCS#1 v1.5 signature.
 #[inline]
-pub(crate) fn verify(
-    pub_key: &RsaPublicKey,
-    prefix: &[u8],
-    hashed: &[u8],
-    sig: &BigUint,
-) -> Result<()> {
-    let hash_len = hashed.len();
-    let t_len = prefix.len() + hashed.len();
-    let k = pub_key.size();
-    if k < t_len + 11 {
-        return Err(Error::Verification);
-    }
-
+fn verify(pub_key: &RsaPublicKey, prefix: &[u8], hashed: &[u8], sig: &BigUint) -> Result<()> {
     let em = uint_to_be_pad(pub_key.raw_int_encryption_primitive(sig)?, pub_key.size())?;
 
-    // EM = 0x00 || 0x01 || PS || 0x00 || T
-    let mut ok = em[0].ct_eq(&0u8);
-    ok &= em[1].ct_eq(&1u8);
-    ok &= em[k - hash_len..k].ct_eq(hashed);
-    ok &= em[k - t_len..k - hash_len].ct_eq(prefix);
-    ok &= em[k - t_len - 1].ct_eq(&0u8);
-
-    for el in em.iter().skip(2).take(k - t_len - 3) {
-        ok &= el.ct_eq(&0xff)
-    }
-
-    if ok.unwrap_u8() != 1 {
-        return Err(Error::Verification);
-    }
-
-    Ok(())
-}
-
-/// prefix = 0x30 <oid_len + 8 + digest_len> 0x30 <oid_len + 4> 0x06 <oid_len> oid 0x05 0x00 0x04 <digest_len>
-#[inline]
-pub(crate) fn generate_prefix<D>() -> Vec<u8>
-where
-    D: Digest + AssociatedOid,
-{
-    let oid = D::OID.as_bytes();
-    let oid_len = oid.len() as u8;
-    let digest_len = <D as Digest>::output_size() as u8;
-    let mut v = vec![
-        0x30,
-        oid_len + 8 + digest_len,
-        0x30,
-        oid_len + 4,
-        0x6,
-        oid_len,
-    ];
-    v.extend_from_slice(oid);
-    v.extend_from_slice(&[0x05, 0x00, 0x04, digest_len]);
-    v
-}
-
-/// Decrypts ciphertext using `priv_key` and blinds the operation if
-/// `rng` is given. It returns one or zero in valid that indicates whether the
-/// plaintext was correctly structured. In either case, the plaintext is
-/// returned in em so that it may be read independently of whether it was valid
-/// in order to maintain constant memory access patterns. If the plaintext was
-/// valid then index contains the index of the original message in em.
-#[inline]
-fn decrypt_inner<R: CryptoRngCore + ?Sized>(
-    rng: Option<&mut R>,
-    priv_key: &RsaPrivateKey,
-    ciphertext: &[u8],
-) -> Result<(u8, Vec<u8>, u32)> {
-    let k = priv_key.size();
-    if k < 11 {
-        return Err(Error::Decryption);
-    }
-
-    let em = priv_key.raw_int_decryption_primitive(rng, &BigUint::from_bytes_be(ciphertext))?;
-    let em = uint_to_zeroizing_be_pad(em, priv_key.size())?;
-
-    let first_byte_is_zero = em[0].ct_eq(&0u8);
-    let second_byte_is_two = em[1].ct_eq(&2u8);
-
-    // The remainder of the plaintext must be a string of non-zero random
-    // octets, followed by a 0, followed by the message.
-    //   looking_for_index: 1 iff we are still looking for the zero.
-    //   index: the offset of the first zero byte.
-    let mut looking_for_index = 1u8;
-    let mut index = 0u32;
-
-    for (i, el) in em.iter().enumerate().skip(2) {
-        let equals0 = el.ct_eq(&0u8);
-        index.conditional_assign(&(i as u32), Choice::from(looking_for_index) & equals0);
-        looking_for_index.conditional_assign(&0u8, equals0);
-    }
-
-    // The PS padding must be at least 8 bytes long, and it starts two
-    // bytes into em.
-    // TODO: WARNING: THIS MUST BE CONSTANT TIME CHECK:
-    // Ref: https://github.com/dalek-cryptography/subtle/issues/20
-    // This is currently copy & paste from the constant time impl in
-    // go, but very likely not sufficient.
-    let valid_ps = Choice::from((((2i32 + 8i32 - index as i32 - 1i32) >> 31) & 1) as u8);
-    let valid =
-        first_byte_is_zero & second_byte_is_two & Choice::from(!looking_for_index & 1) & valid_ps;
-    index = u32::conditional_select(&0, &(index + 1), valid);
-
-    Ok((valid.unwrap_u8(), em, index))
-}
-
-/// Fills the provided slice with random values, which are guaranteed
-/// to not be zero.
-#[inline]
-fn non_zero_random_bytes<R: CryptoRngCore + ?Sized>(rng: &mut R, data: &mut [u8]) {
-    rng.fill_bytes(data);
-
-    for el in data {
-        if *el == 0u8 {
-            // TODO: break after a certain amount of time
-            while *el == 0u8 {
-                rng.fill_bytes(core::slice::from_mut(el));
-            }
-        }
-    }
+    pkcs1v15_sign_unpad(prefix, hashed, &em, pub_key.size())
 }
 
 /// Signing key for PKCS#1 v1.5 signatures as described in [RFC8017 § 8.2].
@@ -496,7 +355,7 @@ where
     pub fn new(key: RsaPrivateKey) -> Self {
         Self {
             inner: key,
-            prefix: generate_prefix::<D>(),
+            prefix: pkcs1v15_generate_prefix::<D>(),
             phantom: Default::default(),
         }
     }
@@ -505,7 +364,7 @@ where
     pub fn random<R: CryptoRngCore + ?Sized>(rng: &mut R, bit_size: usize) -> Result<Self> {
         Ok(Self {
             inner: RsaPrivateKey::new(rng, bit_size)?,
-            prefix: generate_prefix::<D>(),
+            prefix: pkcs1v15_generate_prefix::<D>(),
             phantom: Default::default(),
         })
     }
@@ -700,7 +559,7 @@ where
     pub fn new(key: RsaPublicKey) -> Self {
         Self {
             inner: key,
-            prefix: generate_prefix::<D>(),
+            prefix: pkcs1v15_generate_prefix::<D>(),
             phantom: Default::default(),
         }
     }
@@ -906,18 +765,6 @@ mod tests {
     use signature::{RandomizedSigner, Signer, Verifier};
 
     use crate::{PublicKeyParts, RsaPrivateKey, RsaPublicKey};
-
-    #[test]
-    fn test_non_zero_bytes() {
-        for _ in 0..10 {
-            let mut rng = ChaCha8Rng::from_seed([42; 32]);
-            let mut b = vec![0u8; 512];
-            non_zero_random_bytes(&mut rng, &mut b);
-            for el in &b {
-                assert_ne!(*el, 0u8);
-            }
-        }
-    }
 
     fn get_private_key() -> RsaPrivateKey {
         // In order to generate new test vectors you'll need the PEM form of this key:
